+++ pages/student/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <student-dashboard nav="dashboard">
+  <dashboardsStudent-dashboard nav="dashboard">
     <div class="py-6">
       <div class="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
         <h1 class="text-2xl font-semibold text-gray-900">
@@ -14,7 +14,7 @@
         <!-- /End replace -->
       </div>
     </div>
-  </student-dashboard>
+  </dashboardsStudent-dashboard>
 </template>
 
 <script>