+++ prisma/seed.js
@@ -1,0 +1,23 @@
+/* eslint-disable */
+const { PrismaClient } = require("@prisma/client");
+const prisma = new PrismaClient();
+
+async function main () {
+  await prisma.closedText.create({
+    data: {
+      title: "Allgemeine Fragen zu Thor",
+      text: "Dies ist ein Text mit $$ Lücken. Die Lücken müssen vorher $$ werden.",
+      answers: "diversen,befüllt"
+    }
+  });
+}
+
+main()
+  .catch((e) => {
+    console.error(e);
+    process.exit(1);
+  })
+  .finally(async () => {
+    await prisma.$disconnect();
+  });
+/* eslint-enable */